refactor(kfx): isolate mutation authority planning
