+++ framework/core/architecture/TARGETS.cmake
@@ -89,6 +89,7 @@ set(KUNGFU_EXTENSION_SERVICES_SOURCE_FILES
   "${PROJECT_SOURCE_DIR}/src/runtime/action/action_runtime.cpp"
   "${PROJECT_SOURCE_DIR}/src/runtime/action/domain_profile.cpp"
   "${PROJECT_SOURCE_DIR}/src/runtime/action/profile_action.cpp"
+  "${PROJECT_SOURCE_DIR}/src/runtime/kfx/native_authority.cpp"
   "${PROJECT_SOURCE_DIR}/src/runtime/kfx/native_contract.cpp"
   "${PROJECT_SOURCE_DIR}/src/runtime/kfx/native_registry.cpp"
   "${PROJECT_SOURCE_DIR}/src/runtime/profile/profile_lifecycle.cpp"
